# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DeleteAccountRequest": {
        "properties": {
          "password": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "title": "DeleteAccountRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PushTokenUpdate": {
        "properties": {
          "expo_push_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expo Push Token"
          }
        },
        "title": "PushTokenUpdate",
        "type": "object"
      },
      "QuotaRead": {
        "properties": {
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining"
          },
          "used": {
            "title": "Used",
            "type": "integer"
          }
        },
        "required": [
          "used",
          "limit",
          "remaining"
        ],
        "title": "QuotaRead",
        "type": "object"
      },
...
      "UsageRead": {
        "description": "Günlük kota durumu (bkz. services/usage_quota_service.py).",
        "properties": {
          "chat": {
            "$ref": "#/components/schemas/QuotaRead"
          },
          "photo": {
            "$ref": "#/components/schemas/QuotaRead"
          },
          "resets_at": {
            "format": "date-time",
            "title": "Resets At",
            "type": "string"
          }
        },
        "required": [
          "chat",
          "photo",
          "resets_at"
        ],
        "title": "UsageRead",
        "type": "object"
      },
...
      "UserRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "health_data_consent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Data Consent At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kvkk_consent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kvkk Consent At"
          },
          "terms_consent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Consent At"
          }
        },
        "required": [
          "id",
          "email",
          "created_at"
        ],
        "title": "UserRead",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "PulseCoach API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/users/me": {
      "delete": {
        "description": "Hesabı ve sahip olduğu TÜM veriyi kalıcı olarak siler (User modelindeki\ncascade=\"all, delete-orphan\" relationship'leri sayesinde profil/öğün/\nilerleme/sohbet/mood/checkin/egzersiz/refresh-token'lar da otomatik\nsilinir). Geri alınamaz bir işlem olduğu için mevcut şifrenin tekrar\ngirilmesi zorunlu - ele geçirilmiş bir oturumun tek bir istekle hesabı\nsilmesini engeller.",
        "operationId": "delete_current_user_users_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Current User",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "read_current_user_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Current User",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/export": {
      "get": {
        "description": "Kullanıcının sahip olduğu tüm veriyi (GDPR-tarzı \"verini indir\") tek\nbir JSON belgesi olarak döndürür.",
        "operationId": "export_current_user_data_users_me_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Current User Data",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/push-token": {
      "post": {
        "description": "Expo push token'ını kaydeder/temizler - expo_push_token User\ntablosunda yaşıyor (bir profil tercihi değil, cihaz kaydı), bu yüzden\n/profile PATCH'e KATILMIYOR, users_router'ın diğer User-tablosu/cihaz-\nhesabı endpoint'leriyle (export, delete) aynı yerde.",
        "operationId": "update_push_token_users_me_push_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTokenUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Push Token",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/usage": {
      "get": {
        "description": "Bugünkü sohbet/fotoğraf kullanımı ve kalan hak - istemciler hak azalınca\nuyarı göstermek için okur.",
        "operationId": "read_usage_users_me_usage_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Usage",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}